backend/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
pymongo
python-multipart
python-dotenv
resemblyzer
pydub
numpy
